Move discount badge from ui/ to product/
Domain-specific component (displays "% OFF") doesn't belong in ui/
primitives. Only used by product-card and PDP price components.
