+++ apps/template/components/product/discount-badge.tsx
